--- GameClock/Utility/Items.swift
@@ -1,33 +1,0 @@
-//
-//  Items.swift
-//  GameClock
-//
-//  Created by Ipo Abe on 2022/05/27.
-//
-
-import Foundation
-import UIKit
-
-final class Items {
-    
-    let p1TimeKey = "p1TimeKey"
-    let p2TimeKey = "p2TimeKey"
-    
-    func convertHMS(_ time: Int) -> String {
-        var stringTime = "00:00:00"
-        let hour = time / 3600
-        let min = time % 3600 / 60
-        let sec = time % 3600 % 60
-        
-        let stringHour = String(hour)
-        let stringMin = String(format: "%02d", min)
-        let stringSec = String(format: "%02d", sec)
-        
-        switch time {
-        case (0..<60): stringTime = "\(stringSec)"
-        case (60..<3600): stringTime = "\(stringMin):\(stringSec)"
-        default : stringTime = "\(stringHour):\(stringMin):\(stringSec)"
-        }
-        return stringTime
-    }
-}
